+++ build_runner/lib/src/entrypoint/options.dart
@@ -31,6 +31,7 @@ const releaseOption = 'release';
 const trackPerformanceOption = 'track-performance';
 const skipBuildScriptCheckOption = 'skip-build-script-check';
 const symlinkOption = 'symlink';
+const useExperimentalResolverOption = 'use-experimental-resolver';
 const usePollingWatcherOption = 'use-polling-watcher';
 const verboseOption = 'verbose';
 
@@ -86,6 +87,8 @@ class SharedOptions {
 
   final List<String> enableExperiments;
 
+  final bool useExperimentalResolver;
+
   SharedOptions._({
     required this.buildFilters,
     required this.deleteFilesByDefault,
@@ -100,6 +103,7 @@ class SharedOptions {
     required this.isReleaseBuild,
     required this.logPerformanceDir,
     required this.enableExperiments,
+    required this.useExperimentalResolver,
   });
 
   SharedOptions.fromParsedArgs(ArgResults argResults,
@@ -122,6 +126,8 @@ class SharedOptions {
           isReleaseBuild: argResults[releaseOption] as bool,
           logPerformanceDir: argResults[logPerformanceOption] as String?,
           enableExperiments: argResults[enableExperimentOption] as List<String>,
+          useExperimentalResolver:
+              argResults[useExperimentalResolverOption] as bool,
         );
 }
 
@@ -147,6 +153,7 @@ class DaemonOptions extends WatchOptions {
     required super.logPerformanceDir,
     required super.usePollingWatcher,
     required super.enableExperiments,
+    required super.useExperimentalResolver,
   }) : super._();
 
   factory DaemonOptions.fromParsedArgs(ArgResults argResults,
@@ -187,6 +194,8 @@ class DaemonOptions extends WatchOptions {
       logPerformanceDir: argResults[logPerformanceOption] as String?,
       usePollingWatcher: argResults[usePollingWatcherOption] as bool,
       enableExperiments: argResults[enableExperimentOption] as List<String>,
+      useExperimentalResolver:
+          argResults[useExperimentalResolverOption] as bool,
     );
   }
 }
@@ -214,6 +223,7 @@ class WatchOptions extends SharedOptions {
     required super.isReleaseBuild,
     required super.logPerformanceDir,
     required super.enableExperiments,
+    required super.useExperimentalResolver,
   }) : super._();
 
   WatchOptions.fromParsedArgs(ArgResults argResults,
@@ -237,6 +247,8 @@ class WatchOptions extends SharedOptions {
           logPerformanceDir: argResults[logPerformanceOption] as String?,
           usePollingWatcher: argResults[usePollingWatcherOption] as bool,
           enableExperiments: argResults[enableExperimentOption] as List<String>,
+          useExperimentalResolver:
+              argResults[useExperimentalResolverOption] as bool,
         );
 }
 
@@ -266,6 +278,7 @@ class ServeOptions extends WatchOptions {
     required super.logPerformanceDir,
     required super.usePollingWatcher,
     required super.enableExperiments,
+    required super.useExperimentalResolver,
   }) : super._();
 
   factory ServeOptions.fromParsedArgs(ArgResults argResults,
@@ -337,6 +350,8 @@ class ServeOptions extends WatchOptions {
       logPerformanceDir: argResults[logPerformanceOption] as String?,
       usePollingWatcher: argResults[usePollingWatcherOption] as bool,
       enableExperiments: argResults[enableExperimentOption] as List<String>,
+      useExperimentalResolver:
+          argResults[useExperimentalResolverOption] as bool,
     );
   }
 }